Replace postgres.js pt 1 (#5430)
This replaces all the usages of postgres.js for normal
(non-subscription) queries with Kysely.

The next PR will be replacing the postgres.js subscription with
[pg-logcial-replication](https://github.com/kibae/pg-logical-replication)
and using a persistent replication slot to avoid the need to reboot user
DOs whenever there is connection flake.

Then after that we can look at

- keeping a buffer of previous replication events so if a user DO misses
a message or two it can request a catch-up rather than having to reboot.
- look at doing persistence in the user DOs and requesting catchups when
starting up so during deploys we can avoid triggering a reboot on every
active user DO.

And after that, new socket connections should be able to serve
relatively up-to-date data immediately and we should have our immediate
perf problems solved for a while.

### Change type

- [x] `other`