Add dotcom checks that will serve for alerting (#5440)
Collect some basic stats which can be used for alerting purposes.

I investigated using checkly for some browser tests but decided against
it. The main is is that it's not trivial to login from test automations
like playwright. Even if you set the username / password for Oauth, you
might get hit by captcha / automation detection. Two ways to get around
that would be:
* Allowing sign in via username + password combo.
* Using some sort of email throwaway / burner service, which would allow
api access to get the sign in codes.

This will use our existing updown.io health checks, but will extend them
with some basic app specific health checks. One downside of this is that
we use CF for collecting these checks, so we might report DB / Clerk
being down, when it's in fact CF that is down. That said I don't think
that's a big problem since if CF is down our app is down for sure.

I have added checks to updown.io, will enable them once this lands in
prod.

### Change type

- [x] `improvement`

---------

Co-authored-by: David Sheldrick <d.j.sheldrick@gmail.com>