switch to pg-logical-replication (#5433)
follow up to #5430 

This PR swaps out our postgres.js subscription with
pg-logical-replication, which allows us to use a persistent replication
slot. This in turn means that when the postgres connection breaks we can
reliably restart the replication slot from where we picked off with
guaranteed consistency, and therefore we won't need to force our user
durable objects to reboot.

In the next PR I will make it so that the replicator keeps a log of
events so user DOs can resume via catch-up instead of doing a full
reboot.

### Change type

- [x] `other`