retry slurping document records if need be (#5460)
There was a hole in the local file slurping logic where it would retry
slurping assets but not the document records themselves. So if the user
switched away from the slurped file before tlsync had a chance to finsih
doing its thing, the data would seem to disappear.

This patches that hole by using the `createSource` column to signify
that an app file should be initialized by slurping from a local file,
rather than an ephemeral property on the TldrawApp instance. That way
when we open such a file we can check whether the
document.meta.slurpPersistenceKey has been added, and if so it means the
records slurp happened successfully.

### Change type

- [x] `bugfix`