Consider `https://localhost` to be development (#5471)
I've been occasionally using https for local development, and it'd be
nice not to have the console warning about our license key not covering
the domain.

An alternative solution could be to add `localhost` to the set of
allowed domains for all license keys (or just for users who use https
during local dev.

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>