Improve / fix layout methods: alignment, distribute, flip, stack. (#5479)
This PR fixes various bugs in our layout methods: align, distribute,
flip, stack, pack, etc.

There are a lot of small changes and fixes in this PR but the main one
is **shape clustering**.

## Shape clustering

Previously, all of our layout options had bugs related to arrows. Our
early solution was to ignore arrows, however this only partially solved
the problem. Often the results would not be anywhere near
expectations—obviously wrong—but there wasn't a clear idea of what would
be correct.

While working on this problem, I had the idea to try treating networks
of bound-and-selected shapes as "clusters" that would move together.
This was specifically to solve a problem that could occur when aligning
a shape that had an arrow bound to it and the other end not bound to
anything.

In this PR, all of our layout functions (except flip) move "clusters"
rather than individual shapes—though most clusters only contain a single
shape.

<img width="553" alt="image"
src="https://github.com/user-attachments/assets/7245689c-80bd-4ae0-85df-d42d00251d44"
/>

This fixes many problems which were previously impossible to solve. A
use can bail out of the behavior by not selecting the arrows, just as
they could before.

## Alignment

Alignment now works correctly with clusters of bound shapes.

![Kapture 2025-02-23 at 09 41
23](https://github.com/user-attachments/assets/de75ae58-17ab-430d-871f-cd8a0f3ad05e)

![Kapture 2025-02-23 at 18 19
47](https://github.com/user-attachments/assets/10cc8841-8ec0-4d98-9872-545edcc1002d)

![Kapture 2025-02-23 at 18 21
57](https://github.com/user-attachments/assets/fee0d432-5242-45dd-a47c-c659326f91fa)

## Distribution

Distribution now works correctly with clusters of bound shapes.

Adjusts the logic for distribution so that the result will **space
shapes out evenly between** the first and last shapes, rather than
spacing out the centers of the shapes.

Fixes a bug when distributing shapes if the same shape constituted the
first and last shape.

Fixes a bug when distributing shapes if the last shape would change the
dimensions of the selection.

![Kapture 2025-02-23 at 09 44
11](https://github.com/user-attachments/assets/9991a149-9eb1-43e7-bc7c-0dcc6f1a087b)

![Kapture 2025-02-23 at 18 24
40](https://github.com/user-attachments/assets/2dd815a1-7e0c-424f-b1d5-5377fbe41ddb)

## Flip

I fixed a minor bug with arrows in flipping.

![Kapture 2025-02-23 at 15 03
04](https://github.com/user-attachments/assets/59a1d7cc-6956-4f1b-a772-a4ae1fd4414c)

## Stretch

Stretch now works with clusters of shapes.

Fixed a limitation with rotated shapes. Previously, stretching did not
support any page-rotated shapes unless their page rotation was a
multiple of PI2. It now supports shapes with any rotation equal to PI /
2 (e.g. 90, 180, 270, or 360 degrees, and so on).

## Pack

Pack not works with clusters of shapes.

![Kapture 2025-02-23 at 17 48
12](https://github.com/user-attachments/assets/2b9a35d4-1387-42db-97d1-80966a3822b0)

### Change type

- [x] `bugfix`

### Test plan

1. Align a shape to an edge where that edge in the shape bounds is
defined by an arrow
2. Distribute overlapping or irregular shapes

- [x] Unit tests

### Release notes

- Fixes several bugs when aligning / flipping / distributing /
stretching / stacking a selection that included with arrows.
- Fixed a bug with distribution with overlapping shapes
- Fixed a bug with distribution that could lead to changed selection.
- Fixed a bug preventing rotated shapes from being stretched.