[botcom] Fix slow export menu in big files (#5435)
This PR fixes an issue where large files or files with GIFs / large
images could freeze when opening the export menu.


https://github.com/user-attachments/assets/af0b21c2-d50b-4106-bef6-7738868f863e

## GIFs

Previously, we were inlining the entire GIF when creating an SVG. We now
only inline the first frame. This was a crashing bug previously.

## Images / asset caching

Previously, we were loading a URL for each image or video shape. We now
cache the result of the previous generation, so that we only load one
image per asset. This is especially important for GIFs because the work
to generate the export image src associated with the asset was
expensive.

## File size

Previously, the image shown in the export panel could be very very large
depending on the size of the canvas.

**Problem**

When we export images, we do so at 100% resolution. For large canvases,
that can result in images which are very large (e.g. 5000x5000px or
larger). Creating and displaying such a large image can be extremely
hard on resources.

**Solution**

In this PR, we find a **scale** for the image based on the common bounds
of the selected shapes, then apply that scale to keep the image small
(under 500x500 pixels). We also use png instead of svg as the output
format, which for small but dense images should work better. Together,
these lead to preview images for even very large files (4000 draw
shapes) being relatively quick. The actual exports are unaffected and
slower.

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Visit a very large file.
2. Open the export menu.

### Release notes

- Fixed a bug with export menu performance.