getImageSize: Fix inaccurate PNG image width height calculation by using exact pixels per meter value (#5509)
![2b3ac5f41943b7f399fabd9803b6e52d999fb1f5](https://github.com/user-attachments/assets/61540617-f304-4d43-8d11-312bac38f4ba)

Previously, if you try to upload this image (1320 x 2868) into tldraw
canvas, and then inspect its asset size, you will realize tldraw
processes the image into (1320 x 2867) which is 1 pixel off in height.

This is caused by `getImageSize` function using an inaccurate pixels per
meter value, after rounding off, we lose 1 pixel in the process.

The fix is to use the exact value of pixels per meter, which is
calculated as so:

- 72 DPI is 72 dots per inch
- 1 inch is 0.0254 meter
- Therefore, pixels per meter for 72 DPI is 72 / 0.0254.

And so we just use 72 / 0.0254 as the pixels per inch value instead of
rounding, which gives us the most accurate result.

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Download this test image above
2. Observe in file explorer that its size should be 1320 x 2868
![CleanShot 2025-02-26 at 23 35
48@2x](https://github.com/user-attachments/assets/4fa94203-de09-439b-ab65-b38168560e13)
3. Upload to tldraw canvas
4. Use inspector and issue this command
`editor.getAsset(editor.getOnlySelectedShape().props.assetId)` and see
that the image size is indeed 1320 x 2868


- [x] Unit tests
- [ ] End to end tests

### Release notes

- Fixed a bug with `getImageSize` using inaccurate pixels per meter
value which leads to discrepancies in calculating width and height in
pixels for PNG images.