Add `AtomMap` & refactor store (#5496)
This diff adds `AtomMap`, a drop-in replacement for `Map` that stores
values in reactive atoms. `AtomMap` replaces the old storage mechanism
for the `Store` - an `Atom<ReadonlyMap<Id, Atom<Record>>>`.

The main benefit (and motivation behind this change) is that
adding/removing a record from an `AtomMap` no longer causes everything
accessing a record from the store to recompute. This cuts down on the
number cache invalidations on unrelated changes pretty dramatically. The
fact that `AtomMap` is a motivation too - it does't have a second use in
this diff, but in #4895, this replaces another adhoc atom-map like
structure in the font manager.

There are a couple of additional changes in this diff that I pulled out
of the rich text/fonts PR:
- `createSelectedComputedCache`, which had no uses throughout the
codebase, has been deleted. There's a new `createCache` method that
accepts a function returning a `computed` which can be used to create
selector caches, or anything more complex you might need.
- `createComputedCache` now accepts an options object instead of a plain
`isEqual` callback so we can have callbacks for record-equality and
result-equality.

### Change type

- [x] `api`

### Release notes

- **BREAKING**. `store.createSelectedComputedCache` has been removed.
Use `store.createCache` and create your own selector `computed` instead.
- **BREAKING**. `createComputerCache` no longer accepts a single
`isEqual` fn as its 3rd argument. Instead, pass in an options object,
with the `isEqual` fn named `areRecordsEqual`. You can now pass
`areResultsEqual`, too.