Stabilize useSync invocation (#5512)
Phil reported an issue 

> Created new page in the "sales" project called Sales Funnel v3

> Noticed it was called "untitled"...and when I clicked on the pages
view the site refreshed and the page had gone.

I reckon this was the same kind of thing that started happening randomly
a couple of months ago where clerk, upon token refresh or something,
would return a new object from the useAuth hook which in turn would
invalidate our useTldrawUser hook, which in turn would invalidate our
useSync invocation and cause the entire editor to remount.

So here I've stabilized the useSync invocation so that it only
invalidates when the file slug changes or the auth status (isSignedIn vs
not) changes, and then we use a useEvent to handle fetching the token.

### Change type

- [x] `other`