Fix file dropdown crashing the app on slow connections (#5510)
Fixes an error that would occur on slow connections. 

The issue was that on a slow connection the editor might take a while to
load, which meant that we didn't have the actions provider set yet [due
to this
logic](https://github.com/tldraw/tldraw/blob/165f4f0749222457f70284e779e7eaf9c5616acd/packages/tldraw/src/lib/ui/context/TldrawUiContextProvider.tsx#L80-L85).

I also discovered that we didn't actually show the file dropdown item
that caused the issue in the first place. The reason was that we didn't
set the `fileOverrides` for `TldrawUiContextProvider`
[here](https://github.com/tldraw/tldraw/blob/1a89d9688c44affc84ed01495be37a082a83cc22/apps/dotcom/client/src/tla/providers/TlaRootProviders.tsx#L116-L125).
This meant that the `save-file-copy` action was not available anyway. We
could not make this work with `TldrawUiMenuActionItem` since there's no
way for us to pass in the file slug to the action, so I reworked this to
use regular handlers that don't go through actions.

(Looks like using the actions approach was meant for the main menu,
there we always download the current file and can use the file slug. But
sidebar has many files, so that approach didn't work)

### Change type

- [x] `bugfix`

### Test plan

1. Use slow 4g network setting.
2. Reload the page.
3. Open the file dropdown as soon as you can.
4. It should not crash the app.