Do hard deletes reactively (#5522)
Before this PR the only path we had available for hard deletes was via
client-side mutations. If we deleted the row manually in the db (e.g.
because a user asked us to wipe their data) it would not result in the
R2/KV artefacts being cleaned up.

Before this PR we also needed to get the publsihedSlug from the row
before deleting it so we could clean up the publishing artefacts.

Now I added the publishedSlug to the file table primary key so we get it
in delete events. Then when we handle those delete events in the file
durable object, it does the cleanup there.

We'd end up doing this during the switch to zero anyway I think.

(note this wasn't totally safe to do before because file deletes may
have happened while the replicator didn't have an open subscription to
Postgres. But now we use a persistent replication slot this should
always work)

### Change type

- [x] `other`