Make collaboration hooks public (#5541)
Makes `usePeerIds` and `usePresence` public so they can be used to
customise collaboration components more easily.

I left the comment that maybe these two hooks should be computed
properties, but if they're made public that may not be needed anymore.
Unless it's better to make those computed properties right now and skip
the public hooks?

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [x] `api`
- [ ] `other`

### Test plan

Check that `usePeerIds` and `usePresence` can be accessed. Maybe an
example?

- [ ] Unit tests
- [ ] End to end tests

### Release notes

- Makes `usePeerIds` and `usePresence` public