Make better use of replicator history (#5532)
After making a couple of deploys with the new backend this morning we
noticed that a large portion of user DOs still ended up doing a full db
reset and this seems to be for two reasons:

1. we only stored about 10 minutes worth of history
2. a lot of users in our active_users table are idle at any given time
(maybe like 70-80% of them) so a lot of them will not have received
updates for more than 10 minutes.

This PR makes it so that if a user hasn't done anything for 5 minutes
but their socket is still open, we execute a noop mutation number bump
so that they receive a new `lsn`.

I also cleaned up the mutation handling logic so that it no longer does
two transactions per mutation. This was only required before because
postgres.js would not dispatch events grouped by transaction, and the
ordering was not maintained. wal2json guarantees changes to be grouped
by transaction with formatVersion: 1 (the default that we use) so as
long as we commit the mutations after handling any state updates it
should avoid data races.

### Change type

- [x] `other`

---------

Co-authored-by: Mitja Bezenšek <mitja.bezensek@gmail.com>