Migrate newsletter form to hubspot (#5557)
Submit newsletter subscribers to hubspot instead of sendgrid.

The reason why we don't use a straight api call to add a contact is that
you [can't set the contact to be a marketing
contact](https://community.hubspot.com/t5/APIs-Integrations/How-do-I-set-a-contact-as-a-marketing-contact-using-Hubspot-s/m-p/895454)
via an API. So I created a form in hubspot and we then [submit to that
form](https://community.hubspot.com/t5/APIs-Integrations/Cannot-set-a-contact-s-marketing-status-via-API/m-p/800944/highlight/true#M64478).

An alternative approach would be to use hubspot's form directly, but
since this form is used in a couple of different places with different
styling options it would make it a bit annoying to correctly style.

### Change type

- [x] `improvement`