Use ImmutableMap inside AtomMap (#5567)
This PR fixes load perf for large files by updating `AtomMap` to use an
immutable map (taken from Immutable.js).

On a large room like our old retro board, populating the `AtomMap` would
take up more than a second in local dev on my machine, this brings it
down to < 15ms. The cause for the slowness was updating the presentKeys
`Set` via copy-on-write.

This also fixes another issue with AtomMap in that it didn't respect
signia transaction rollbacks because it used a mutable Map to store the
atoms.

### Change type

- [x] `bugfix`

### Release notes

- Fixed a perf regression that caused slowness mainly when loading
documents.