Add more postgres history, reduce ping overhead (#5604)
Last time we did a dotcom release during working hours there were way
more User DO full postgres refetches than anticipated, which led to a
couple of minutes of slowness.

Investigating this, I couldn't find a link in the chain that had failed,
so it seems likely that the history pruning was too aggressive and the
length of time between 'noop' mutations was too long. So in this PR I
tweak both of those variables, and make 'noop' mutations more efficient
by bypassing postgres and just asking the replicator for an lsn update
directly.

If this doesn't work there is definitely something I'm missing about the
production env that makes this stuff fail. In this PR I also added a
bunch of extra analytics to maybe help narrow it down after the next
release.

### Change type

- [x] `other`