[Hotfix] dismissibles (#5659)
…570)

We used to default to "user didn't dismiss" this yet. Which meant that
we always showed a dismissible element first, then we read the actual
value, and we then might have dismissed the elements (if that's what the
stored state dictated). We now have a "not initialized" state during
which we don't show the elements.

Side effect of this change is that some users that have already
dismissed these elements might see them again. I guess we could add some
sort of migration of the state, but meh, doesn't seem worth the hassle?

- [x] `improvement`

Co-authored-by: Mitja Bezenšek <mitja.bezensek@gmail.com>