assets: be able to remove assets that are unused (#5628)
Be able to cleanup the assets when the assets are actually deleted. 
This came up in Discord here:
https://discord.com/channels/859816885297741824/1349476132365537320/1349476132365537320

The IndexedDB can start getting crufty with old assets that are unused.
This allows folks that proactively delete assets (not just the shapes)
to be able to keep the local db clean.

fixes https://github.com/tldraw/tldraw/issues/5622

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Cleanup assets from the local indexedDB that are proactively deleted.