upgrade yarn to 4.7 (#5687)
Upgrade yarn to 4.7. The version of 4 we were using caused some issues
when testing react 19.

If downstream users are using corepack, the new version of yarn should
be fetched and used automatically the first time they use `yarn`.

4.7 has some stricter rules around peer dependencies. I've fixed the
glaring ones, but we now get 'completed with warnings' due to issues in
our dependencies dependencies :( I've disabled the no-warnings check as
a result.

### Change type

- [x] `other`