[Fix] Rich text perf issue (#5658)
This PR improves the performance of the app with regard to text.


https://github.com/user-attachments/assets/3c468587-0cc1-4e72-8979-61fe14f6a0dc

The `renderPlaintextFromRichText` method is slow. It depends on TipTap's
`getText` method. This parses the rich text and flattens it to plain
text. As usual, we can improve performance by:

1. making `renderPlaintextFromRichText` faster
2. not calling `renderPlaintextFromRichText`

## Making it faster 

We can make it faster by using a weak cache so that we don't need to
re-flatten text if we've flattened it before. We can also make it faster
by skipping the flatten if the text content is empty. This can be done
very quickly without using `getText`.

## Avoiding it entirely

It turns out that most of the places where we call
`renderPlaintextFromRichText` are checks against empty text, meaning we
don't need to call `renderPlaintextFromRichText` at all. In these cases
we _just_ check whether the text is empty and continue based on that
result.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Improved performance related to rich text.

---------

Co-authored-by: Mime Čuvalo <mimecuvalo@gmail.com>