Use pg message bus for requesting lsn update (#5722)
After the deploy this morning we started getting a lot of 

> Error: Subrequest depth limit exceeded. This request recursed through
Workers too many times. This can happen e.g. if you have a Worker or
Durable Object that calls other Workers or objects recursively.

In sentry. All seeming to originate from the new RPC-based lsn update
mechanism. No idea why, there's a back and forth but no real chain of
recursion here, it's just running on a setInterval.

Anyway this cuts one side of the RPC link out so it can't possibly think
there's a recursion chain.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`