When editing a text shape, don't mount the text editor for non-editing empty shapes unless they're hovered (#5734)
This PR improves text editing.

## The bug

Previously, we would mount all of the text editors for all of the
text-editable shapes when the user began editing any text-editable
shape. For large projects this would block the main thread as all of
those text editors were mounted. Even on my Very Fast Laptop this would
introduce ~200-500ms of lag between when I press "Enter" to begin
editing a shape and when the app would begin responding to my key
strokes. This often dropped the first letter of whatever I was typing.

## The fix

We now mount the text editor for a shape if it's:
- editing; or
- has text; or
- is hovered

If a shape is neither of those, then we don't mount the text editor
until the user hovers the shape. From a user perspective, the
interactions are identical to what they were before.

## Proof

Before: (starting to edit a shape with ~1000 empty shapes on the page)

<img width="1334" alt="image"
src="https://github.com/user-attachments/assets/820908a2-21f0-4a3e-a5b9-5083e4ef6d66"
/>

After:

There's nothing to see, we don't drop any frames.

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Create many empty shapes
2. Begin editing any shape.

### Release notes

- Fixed a bug causing a performance delay when editing text.

---------

Co-authored-by: Mime Čuvalo <mimecuvalo@gmail.com>