Cache the fonts extracted from rich text. (#5735)
This PR improves the performance of shapes with rich text content.

## The bug

Due to reasons I cannot understand, we extract the fonts from rich text
on every frame while translating shapes. Maybe we shouldn't, I don't
know. Anyway, this is expensive when many shapes are moving around.
Actually, not even that many. Try it yourself!

## The fix

The fix here is the same as #5658, which is to skip the work if the text
is empt; or use a weak cache to re-use the result from last time.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Create lots of geo shapes.
2. Drag all the geo shapes around.

### Release notes

- Improved performance while editing many geo shapes or text shapes.

---------

Co-authored-by: Mime Čuvalo <mimecuvalo@gmail.com>