a11y: add a live region to announce selected tools (#5634)
Starts improving our voiceover, starting with the toolbar which was
crap.
This follows the toasts pattern we have, since this is, in a way, a
notification of what's happening on the screen, especially when
selecting something using a hotkey instead of using the mouse pointer.


https://github.com/user-attachments/assets/95f7551a-e891-4c5a-bf93-784b86fe618f

related issue https://github.com/tldraw/tldraw/issues/5215


### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Adds better voiceover support when selecting an action (a11y)