fix zoom speed for pinch gestures (#5771)
Alternative take on #5749, aiming to fix #4945. 

The big issue here is that `useGesture` doesn't give us a zoom delta, it
gives us its own calculated absolute zoom. This makes it really hard to
apply `zoomSpeed` correctly - we don't have a delta to apply a speed to,
and `useGesture` isn't expecting us to be applying changes different to
the ones that it's supplying.

In #5749, we try to get the deltas out and scale them appropriately. In
this diff, we instead counter-scale all of our inputs to `useGesture`,
then scale them back afterwards. This means that if you have a zoom
speed of 0.5, then scaling from 1 to 2 requires scaling from 1 to 4 in
use-gesture land.

The big drawback to this is that where the rest of `cameraOptions` is
applied in editor-land, this is applied to the events _before_ they get
to the editor. This isn't really a big deal - i doubt many people are
using the editor without the built-in use-gesture setup. In the future
we'll probably move off of use-gesture and can apply this stuff in the
editor once we do that.

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Setting `zoomSpeed` in camera options no longer breaks zooming on
safari trackpads and multitouch pinch to zoom.