Prune sst stages based on the active amazon clusters. (#5805)
Gets active clusters from aws (sst does not have a list command), then
we prune sst stages for prs that have been closed for a while.

Ran this locally and got the arns and then the stage names correctly.
Let's see if the auth is set up correctly when this next runs.

```
💡[424]: test.ts:10: arn= arn:aws:ecs:eu-north-1:571248596162:cluster/tldraw-pr-5551-clusterCluster-trnfsoeh
💡[424]: test.ts:10: arn= arn:aws:ecs:eu-north-1:571248596162:cluster/tldraw-davidsheldrick-clusterCluster-xtawvuow
💡[424]: test.ts:10: arn= arn:aws:ecs:eu-north-1:571248596162:cluster/tldraw-pr-5804-clusterCluster-fhmtdafa
💡[423]: test.ts:16: names= [ 'pr-5551', 'pr-5804' ]
```

Resolves INT-1245
(together with changes in #5795)

### Change type

- [x] `improvement`