isShapeHidden => getShapeVisibility, to allow children of hidden shapes to be visible (#5762)
This PR was motivated by a very reasonable [discord request

](https://discord.com/channels/859816885297741824/1353628236487327857/1353628236487327857)

> We are working on a feature where we need to hide all of the shapes
except for the "focused shape" (we are using isShapeHidden prop) - this
is largely working as expected, except for one challenge - when the
"focused shape" is child of a frame/ group shape, then hiding the parent
shape also hides the "focused shape", which makes sense in general
context, but we need the ability to hide the parent shape without hiding
the "focused shape".

Ended up being a fairly small diff.

![Kapture 2025-03-27 at 12 00
09](https://github.com/user-attachments/assets/e2423d49-9908-4a7b-bdb0-fed96c3b25f5)

I'm not crazy about this 'force_show' API and would appreciate
alternative suggestions if you have any.

### Change type

- [x] `other`

### Release notes

- Allow the children of a hidden shape to show themselves by returning a
'force_show' override from the `isShapeHidden` predicate.