a11y: navigable shapes (#5761)
As part of a [larger push](https://github.com/tldraw/tldraw/issues/5215)
to add accessibility to our SDK, a big piece of that work is being able
to navigate through our shapes in some kind of predictable fashion. This
builds upon @Taha-Hassan-Git 's great work and knowledge in this area,
thanks man. :tip-o-the-hat:

Things that were tackled in this PR:
- navigating shapes using the Tab key, when in the Select tool.
- navigating shapes using Cmd/Ctrl+Arrow keys, when in the Select tool.
- only allowing certain shapes to be navigated to. We ignore
draw/highlighter/arrow/group/line. Groups need exploration and will be
tackled later.
- panning the camera to the selected shape, but avoiding doing so in a
jarring way. We don't center the shape to avoid too much whiplashy-ness.

An initial foray into this was relaying purely on DOM but it had a bunch
of browser quirks which forced making this purely a programmatic control
on our end. Things like ensuring culled shapes are still accessible even
though they're not rendered was one of the issues but also tab order
became unpredictable at times which steered me away from that direction.

We coud have considered using something like rbush for some spatial
indexing of the shapes. For the intents and purposes of this PR, it
seemed like overkill at the moment. But we might cross that bridge down
the line, we'll see.

The reading-direction heuristics are a combination of dividing the pages
into rows and then looking at distance and angles to see what is the
spatially "next" shape to be read. It takes _all_ of the shapes and
sorts them into a logical order so that nothing is missed/skipped when
tabbing around.
The directional-arrow heuristics don't divide things into rows and don't
create a sorted set of shapes. Instead, they decide based on the current
shape and direction which is the next spatially to go to, depending on
distance+angle.

There's a decent amount of nuance in this kind of navigation but it's
not all covered in this PR, for separate PRs, we'll look at:
- [x] adding a "skipping to content" button
- [ ] question whether maybe directional navigation visits ‘canTabTo’
shapes, maybe yes?
- [ ] tackling what Enter/Escape should do when on the canvas shapes
- [ ] how to deal with hierarchy / parent-child / frame / group shapes
- [ ] and more



https://github.com/user-attachments/assets/49b6b34e-2553-4047-846f-5d3383e1e3c6


### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [x] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

- [x] Unit tests
- [x] End to end tests

### Release notes

- a11y: navigable shapes using Tab and Cmd/Ctrl+Arrow